+++ lib/utils/open_link.dart
@@ -1,4 +1,6 @@
 import 'dart:io';
+
+import 'package:updat/updat_exception.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
@@ -6,7 +8,7 @@ Future<void> openLink(String url) async {
   if (await canLaunchUrlString(url)) {
     await launchUrlString(url);
   } else {
-    throw "Error";
+    throw UpdatException('Unable to open link: $url');
   }
 }
 
@@ -14,7 +16,7 @@ Future<void> openUri(Uri uri) async {
   if (await canLaunchUrl(uri)) {
     await launchUrl(uri);
   } else {
-    throw "Error";
+    throw UpdatException('Unable to open URI: $uri');
   }
 }
 
@@ -32,18 +34,17 @@ Future<void> openCustom(
   String path, {
   String? customLocation,
 }) async {
-  if (customLocation != null) {
-    return await openPath(path);
+  if (customLocation == null) {
+    return openPath(path);
   }
   if (Platform.isMacOS) {
     await Process.run(
-      "open",
-      ['-a $customLocation', path],
-      runInShell: true,
+      'open',
+      ['-a', customLocation, path],
     );
   } else {
     await Process.run(
-      customLocation!,
+      customLocation,
       [path],
       runInShell: true,
     );